FEAT: add devices configuration
Added a new page that is available when all the sensors are connected. This page allows the user to associate each wearable to the corresponding position on the body. This is done by calculating the standard deviation of the accelerometer data of each sensor. The one with the highest value is the one that the user moved and is therefore associated to the position selected by the user.