FIX: user's sex info retrieval
Before, each time the sensor configuration buttons were pressed, the widget.sexController.text value would get reset because the widget would be rebuilt. Now the default value for the dropdown button is stored in a different variable and a validation check was added to be sure the user selected the sex.